Cost coins for building tower

+++ GameProject/src/main/java/Models/Towers/ArcherTower.java
@@ -3,7 +3,7 @@ package Models.Towers;
 import Models.Towers.Tower;
 
 public class ArcherTower extends Tower {
-    public ArcherTower(int destroyPower, long bulidCost, int range) {
+    public ArcherTower(int destroyPower, int bulidCost, int range) {
         super(destroyPower, bulidCost, range,"Photos/ArcherTower.png");
     }
 }
