+++ GameProject/src/main/java/Models/Towers/WizardTower.java
@@ -3,7 +3,7 @@ package Models.Towers;
 import Models.Towers.Tower;
 
 public class WizardTower extends Tower {
-    public WizardTower(int destroyPower, long bulidCost, int range) {
+    public WizardTower(int destroyPower, int bulidCost, int range) {
         super(destroyPower, bulidCost, range,"WizardTower.png");
     }
 }
